# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  api.BuildArtifactResponse:
    properties:
      artifact_type:
        type: string
      build_id:
        type: string
      checksum_sha256:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      download_url_path:
        type: string
      id:
        type: string
      name:
        type: string
      path:
        type: string
      size_bytes:
        type: integer
      step_id:
        type: string
      storage_provider:
        type: string
      version_tags:
        items:
          $ref: '#/definitions/api.VersionTagResponse'
        type: array
    type: object
  api.BuildArtifactsEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.BuildArtifactsResponse'
    type: object
  api.BuildArtifactsResponse:
    properties:
      artifacts:
        items:
          $ref: '#/definitions/api.BuildArtifactResponse'
        type: array
      build_id:
        type: string
    type: object
  api.BuildEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.BuildResponse'
    type: object
  api.BuildListEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.BuildListResponse'
    type: object
  api.BuildListResponse:
    properties:
      builds:
        items:
          $ref: '#/definitions/api.BuildResponse'
        type: array
    type: object
...
  api.BuildResponse:
    properties:
      actor:
        type: string
      attempt_number:
        type: integer
      build_number:
        type: integer
      created_at:
        type: string
      current_step_index:
        type: integer
      delivery_id:
        type: string
      error_message:
        type: string
      event_type:
        type: string
      finished_at:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/api.ImageExecutionResponse'
      job_id:
        type: string
      pipeline_config_yaml:
        type: string
      pipeline_name:
        type: string
      pipeline_path:
        type: string
      pipeline_source:
        type: string
      priority:
        type: integer
      project_id:
        type: string
      project_name:
        type: string
      project_slug:
        type: string
      queued_at:
        type: string
      ref_type:
        type: string
      repository_name:
        type: string
      repository_owner:
        type: string
      repository_url:
        type: string
      rerun_from_step_index:
        type: integer
      rerun_of_build_id:
        type: string
      scm_provider:
        type: string
      source:
        $ref: '#/definitions/api.BuildSourceResponse'
      source_author_email:
        type: string
      source_author_name:
        type: string
      source_commit_sha:
        type: string
      source_committer_email:
        type: string
      source_committer_name:
        type: string
      source_ref:
        type: string
      source_sha:
        type: string
      started_at:
        type: string
      status:
        type: string
      trigger_commit_sha:
        type: string
      trigger_kind:
        type: string
      trigger_ref:
        type: string
      trigger_type:
        type: string
      triggered_by:
        type: string
    type: object
  api.BuildSourceInput:
    properties:
      commit_sha:
        type: string
      ref:
        type: string
      repository_url:
        type: string
    type: object
  api.BuildSourceResponse:
    properties:
      ref:
        type: string
      repository_url:
        type: string
      source_commit_sha:
        type: string
    type: object
...
  api.CreateBuildRequest:
    properties:
      project_id:
        type: string
      source:
        $ref: '#/definitions/api.BuildSourceInput'
      steps:
        items:
          $ref: '#/definitions/api.CreateBuildStepInput'
        type: array
      template:
        type: string
    type: object
  api.CreateBuildStepInput:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      timeout_seconds:
        type: integer
      working_dir:
        type: string
    type: object
...
  api.CreatePipelineBuildRequest:
    properties:
      pipeline_yaml:
        type: string
      project_id:
        type: string
      source:
        $ref: '#/definitions/api.BuildSourceInput'
    type: object
...
  api.CreateRepoBuildRequest:
    properties:
      commit_sha:
        type: string
      pipeline_path:
        type: string
      project_id:
        type: string
      ref:
        type: string
      repo_url:
        type: string
    type: object
...
  api.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorBody'
    type: object
...
  api.ImageExecutionResponse:
    properties:
      managed_image_id:
        type: string
      managed_image_version_id:
        type: string
      requested_ref:
        type: string
      resolved_ref:
        type: string
      source_kind:
        type: string
      version_tags:
        items:
          $ref: '#/definitions/api.VersionTagResponse'
        type: array
    type: object
...
  api.VersionTagResponse:
    properties:
      artifact_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      job_id:
        type: string
      kind:
        type: string
      managed_image_version_id:
        type: string
      target_type:
        type: string
      version:
        type: string
    type: object
info:
  contact: {}
  description: HTTP API for Coyote CI control-plane workflows.
  title: Coyote CI API
  version: "0.1"
paths:
...
  /builds:
    get:
      description: Lists builds sorted by newest first with optional pagination and
        project filters.
      parameters:
      - description: Max results (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Number of results to skip
        in: query
        name: offset
        type: integer
      - description: Filter builds by project id
        in: query
        name: project_id
        type: string
      - description: Filter builds by project slug
        in: query
        name: project_slug
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildListEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List builds
      tags:
      - builds
    post:
      consumes:
      - application/json
      description: Creates a new build in pending status.
      parameters:
      - description: Build create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateBuildRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.BuildEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create build
      tags:
      - builds
...
  /builds/{buildID}/artifacts:
    get:
      description: Returns persisted artifact metadata for a build.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildArtifactsEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List build artifacts
      tags:
      - builds
  /builds/{buildID}/artifacts/{artifactID}/download:
    get:
      description: Streams stored artifact content for a build artifact.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      - description: Artifact ID
        in: path
        name: artifactID
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: binary payload
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Download build artifact
      tags:
      - builds
...
  /builds/pipeline:
    post:
      consumes:
      - application/json
      description: Parses and validates pipeline YAML, then creates a queued build
        with resolved steps.
      parameters:
      - description: Pipeline build create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreatePipelineBuildRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.BuildEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create build from pipeline YAML
      tags:
      - builds
  /builds/repo:
    post:
      consumes:
      - application/json
      description: Clones a repository, loads .coyote/pipeline.yml, then creates a
        queued build.
      parameters:
      - description: Repo build create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateRepoBuildRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.BuildEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create build from repository
      tags:
      - builds
...
schemes:
- http
swagger: "2.0"